simplify edax handling code
